requirements.txt
google-adk[database]==1.4.2
python-dotenv==1.1.0



